fixing

+++ src/main/java/ru/yandex/practicum/filmorate/storage/db/UsersLikesFilmsDbStorage.java
@@ -1,23 +1,26 @@
 package ru.yandex.practicum.filmorate.storage.db;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Primary;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
+import ru.yandex.practicum.filmorate.exception.InsertDbException;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.storage.model.UsersLikesFilmsStorage;
+
+import java.time.LocalDateTime;
 
 @Repository
 @Qualifier("usersLikesFilmsDbStorage")
 @Primary
 @Slf4j
+@RequiredArgsConstructor
 public class UsersLikesFilmsDbStorage implements UsersLikesFilmsStorage {
+
     private final JdbcTemplate jdbcTemplate;
 
-    public UsersLikesFilmsDbStorage(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
 
     @Override
     public int getLikesCount(long filmId) {
@@ -33,6 +36,15 @@ public class UsersLikesFilmsDbStorage implements UsersLikesFilmsStorage {
     @Override
     public void addLikeFilm(long filmId, long userId) {
         try {
+            // Записываем событие по добавлению лайка в БД
+            LocalDateTime timeOperation = LocalDateTime.now();
+            String sqlAddEvent = "INSERT INTO user_events(timestamp, user_id, event_type, operation, entity_id) " +
+                    "VALUES (?, ?, ?, ?, ?)";
+            int countUpd = jdbcTemplate.update(sqlAddEvent, timeOperation, userId, "LIKE", "ADD", filmId);
+            if(countUpd != 1){
+                throw new InsertDbException("Ошибка при записи в БД события по добавлению лайка");
+            }
+            // Выполняем добавление лайка.
             String sql = "INSERT INTO users_likes_films(film_id, user_id) VALUES (?, ?)";
             jdbcTemplate.update(sql, filmId, userId);
         } catch (Exception e) {
@@ -44,6 +56,15 @@ public class UsersLikesFilmsDbStorage implements UsersLikesFilmsStorage {
     @Override
     public void deleteLikeFilm(long filmId, long userId) {
         try {
+            // Записываем событие по удалению лайка в БД
+            LocalDateTime timeOperation = LocalDateTime.now();
+            String sqlAddEvent = "INSERT INTO user_events(timestamp, user_id, event_type, operation, entity_id) " +
+                    "VALUES (?, ?, ?, ?, ?)";
+            int countUpd = jdbcTemplate.update(sqlAddEvent, timeOperation, userId, "LIKE", "REMOVE", filmId);
+            if(countUpd != 1){
+                throw new InsertDbException("Ошибка при записи в БД события по удалению лайка");
+            }
+            // Выполняем удаление лайка.
             String sql = "DELETE FROM users_likes_films WHERE film_id = ? AND user_id = ?";
             int rowsUpdated = jdbcTemplate.update(sql, filmId, userId);
             if (rowsUpdated == 0) {
